fix: remove suffices from android shared libs for boost
Recipe build libs named like liboost_*.so.[version]
which is not supported by Android.
Furthermore, linux convention is used by creating
a symlink with simple lib name that points to
the actual library with a version suffix.
This result, when cross building on Windows for
Android to have library duplication, the symlinks
being replaced by a copy of the library.
This patch seems to fix the issue: produced
libraries names follow the pattern *.so, and only
one instance of each is created.
Warning: some produced libraries seem to differ from last
recipe, and built libraries are not exactly the same:
misses:
fiber, locale

Command used:
conan create . boost/1.74.0@conan-solar/stable --build missing
  -s build_type=Debug -s compiler.cppstd=17 -o shared=True
  -o zlib=False -o bzip2=False -o without_stacktrace=True