fix: propagate shared option to glog (and glfags)
New recipes for these modules have changed default value for 'shared'.
Now, 'shared' is false by default.
This caused an issue when building opencv contrib with shared=True.
The build system looked for shared libs for glflags, didn't find
them and then automatically removed the production of the SFM
module.
But since the recipe did not see this, the package() method
didn't remove sfm from the produced libraries.
This caused the generated compilation/link command lines
to still have sfm, and the linking step failed because the
sfm library was not found.